% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\docType{data}
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz Function}
\format{
An object of class \code{function} of length 1.
}
\usage{
FizzBuzz
}
\arguments{
\item{n}{A real number}
}
\value{
1:n using Fizz Buzz pattern
}
\description{
Function that takes in an integer from the user and produces the appropriate Fizz Buzz pattern.
The function prints out every number from 1 to the user input.
Every number that is divisible by 3 and 5 gets replaced with 'Fizz Buzz',
every number only divisible by 3 gets replaced with 'Fizz' and every number only divisible by 5
is replaced with 'Buzz'
}
\examples{
FizzBuzz(15)
FizzBuzz(30)
}
\keyword{datasets}
