% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather Information from Pulliam Airport in Flagstaff, AZ}
\format{
a data frame with 22967 observations
\describe{
  \item{DATE}{The date of observation as a POSIX date format.}
  \item{TMAX}{Daily maximum temperature in degrees Farhenheit.}
  \item{TMIN}{Daily minimun temperature in degrees Farhenheit.}
  \item{PRCP}{Daily precipitation in inches.}
  \item{SNOW}{Daily snow fall in inches.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
A time series of daily max temperatures,  min temperature, precipitation,
and snow in Flagstaff, AZ.
}
\keyword{datasets}
